use fixture for collection test
