+++ src/Apps/Collect/CollectionAppFixture.ts
@@ -1,0 +1,40 @@
+import { ArtworksGridEdges } from "Components/ArtworkGrid/__stories__/ArtworkGridFixture"
+
+export const CollectionAppFixture = {
+  collection: {
+    id: "5bc89e25ec650e14cf0b4061",
+    slug: "kaws-companions",
+    title: "KAWS: Companions",
+    description:
+      "<p>Brian Donnelly, better known as KAWS, spent the first year of his career as an animator for Disney. After leaving in 1997, KAWS took inspiration from the company&rsquo;s signature cartoon, Mickey Mouse, to create his own set of characters that he named &ldquo;Companions.&rdquo; With gloved hands and X&rsquo;s for eyes, &ldquo;Companions&rdquo; first appeared in KAWS&rsquo;s graffiti works across New York City in the late 1990s. By the end of the decade, the street artist created his first three-dimensional version, and his characters have since taken on a variety of colors, sizes, and poses. In 2017, his four-foot-tall Seated Companion (2011) broke the auction record for the series, selling for over $400,000. However, many of KAWS&rsquo;s &ldquo;Companions&rdquo; are considerably more affordable, such as his vinyl toys produced in collaboration with esteemed manufacturers including Bounty Hunter, Bape, Medicom, and his own brand OriginalFake, which was active between 2006 and 2013.</pheaderI",
+    headerImage:
+      "https://artsy-vanity-files-production.s3.amazonaws.com/images/kaws2.png",
+    category: "Collectible Sculptures",
+    credit: "<p>&copy; KAWS, Medicom Toy 2007.</p>",
+    query: { artist_ids: [], artist_id: null, gene_id: null, __id: null },
+    artworks: {
+      aggregations: [],
+      __id: "123",
+    },
+    filtered_artworks: {
+      __id:
+        "RmlsdGVyQXJ0d29ya3M6eyJhZ2dyZWdhdGlvbnMiOlsidG90YWwiXSwiYXJ0aXN0X2lkcyI6W10sImdlbmVfaWRzIjpbXSwic2l6ZSI6MCwic29ydCI6Ii1wYXJ0bmVyX3VwZGF0ZWRfYXQiLCJ0YWdfaWQiOiJjb21wYW5pb24ifQ==",
+      artworks: {
+        pageInfo: {
+          hasNextPage: true,
+          endCursor: "YXJyYXljb25uZWN0aW9uOjI5",
+        },
+        pageCursors: {
+          around: [
+            { cursor: "YXJyYXljb25uZWN0aW9uOi0x", page: 1, isCurrent: true },
+          ],
+          first: null,
+          last: null,
+          previous: null,
+        },
+        edges: ArtworksGridEdges,
+      },
+    },
+    __id: "5bc89e25ec650e14cf0b4061",
+  },
+}
